United all json helpers into 1 class
