--- Helpers/JsonFileReaderSingle.cs
@@ -1,18 +1,0 @@
-﻿using Newtonsoft.Json;
-using System;
-using System.Collections.Generic;
-using System.IO;
-using System.Linq;
-using System.Threading.Tasks;
-
-namespace CS10_Project_Fisketorvet_V1.Helpers
-{
-    public class JsonFileReaderSingle<T>
-    {
-        public static T ReadJson(string JsonFileName)
-        {
-            string jsonString = File.ReadAllText(JsonFileName);
-            return JsonConvert.DeserializeObject<T>(jsonString);
-        }
-    }
-}
