--- Helpers/JsonFileWritter.cs
@@ -1,20 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.IO;
-using System.Linq;
-using System.Threading.Tasks;
-
-namespace CS10_Project_Fisketorvet_V1.Helpers
-{
-    public class JsonFileWritter<T>
-    {
-        public static void WriteToJson(Dictionary<int, T> objects, string JsonFileName)
-        {
-            string output = Newtonsoft.Json.JsonConvert.SerializeObject(objects,
-                                                               Newtonsoft.Json.Formatting.Indented);
-
-            File.WriteAllText(JsonFileName, output);
-        }
-
-    }
-}
